requirements.txt
# Core ML and Deep Learning
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.7.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.2.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
pandas>=1.4.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Metrics and Evaluation
scikit-image>=0.19.0

# Configuration and Logging
pyyaml>=6.0
tensorboard>=2.10.0
tqdm>=4.64.0

# Medical Imaging (Optional but recommended)
pydicom>=2.3.0
nibabel>=4.0.0

# Development and Testing
pytest>=7.1.0
black>=22.6.0
flake8>=5.0.0

# Jupyter (for experimentation)
jupyter>=1.0.0
ipywidgets>=8.0.0
